Update `dev/make-distrib.sh` to support Ivette plugins

See merge request frama-c/frama-c!5264


Approved-by: Allan Blanchard <allan.blanchard@cea.fr>

+++ dev/make-distrib.sh
@@ -72,7 +72,7 @@ do
             ;;
         "--codename")
             shift
-            VERSION_CODENAME=$1
+            VERSION_CODENAME="$1"
             ;;
         "--ci-link")
             CI_LINK=yes
@@ -89,7 +89,7 @@ done
 # Target Names
 
 VERSION=$(cat VERSION)
-VERSION_SAFE=${VERSION/~/-}
+VERSION_SAFE="${VERSION/~/-}"
 
 FRAMAC="frama-c-$VERSION_SAFE-$VERSION_CODENAME"
 if [ "$USE_STASH" == "yes" ]; then
@@ -114,15 +114,31 @@ fi
 IGNORED_FILES="$(git ls-files --ignored --exclude-standard -c)"
 if [ "" != "$IGNORED_FILES" ]; then
   echo "Some versioned files are ignored by .gitignore:"
-  echo $IGNORED_FILES
+  echo "$IGNORED_FILES"
   exit 2
 fi
 
 ################################################################################
 # External Plugins
 
-PLUGINS=$(find src/plugins -mindepth 1 -maxdepth 1 -type d)
-EXTERNAL_PLUGINS=$(find src/plugins -type d -name ".git" | sed "s/\/.git//")
+# (using declare and a while read loop because MacOS is still on bash 3.2 by
+#  default and does not know readarray D:)
+
+declare -a FC_PLUGINS
+while IFS= read -r -d $'\0' p; do FC_PLUGINS+=("$p"); done < <(
+  find src/plugins -mindepth 1 -maxdepth 1 -type d -print0)
+declare -a IVETTE_PLUGINS
+while IFS= read -r -d $'\0' p; do IVETTE_PLUGINS+=("$p"); done < <(
+  find ivette/src/frama-c/plugins -mindepth 1 -maxdepth 1 -type d -print0)
+declare -a FC_EXTERNAL_PLUGINS
+while IFS= read -r -d $'\0' p; do FC_EXTERNAL_PLUGINS+=("$p"); done < <(
+  find src/plugins -type d -name ".git" -print0 | sed "s/\/.git//g")
+declare -a IVETTE_EXTERNAL_PLUGINS
+while IFS= read -r -d $'\0' p; do IVETTE_EXTERNAL_PLUGINS+=("$p"); done < <(
+  find ivette/src/frama-c/plugins -type d -name ".git" -print0 | sed "s/\/.git//g")
+
+PLUGINS=("${FC_PLUGINS[@]}" "${IVETTE_PLUGINS[@]}")
+EXTERNAL_PLUGINS=("${FC_EXTERNAL_PLUGINS[@]}" "${IVETTE_EXTERNAL_PLUGINS[@]}")
 
 ################################################################################
 # Summary
@@ -130,15 +146,26 @@ EXTERNAL_PLUGINS=$(find src/plugins -type d -name ".git" | sed "s/\/.git//")
 echo "----------------------------------------------------------------"
 echo "Make Distribution"
 echo "Version: $VERSION ($VERSION_CODENAME)"
-echo "Plugins: $EXTERNAL_PLUGINS"
+echo "Frama-C Plug-ins:"
+if [ "${#FC_EXTERNAL_PLUGINS[@]}" -gt 0 ]; then
+  printf " * %s\n" "${FC_EXTERNAL_PLUGINS[@]}"
+fi
+echo "Ivette Plug-ins:"
+if [ "${#IVETTE_EXTERNAL_PLUGINS[@]}" -gt 0 ]; then
+  printf " * %s\n" "${IVETTE_EXTERNAL_PLUGINS[@]}"
+fi
 echo "----------------------------------------------------------------"
 
 ################################################################################
 # Warn if there are uncommitted changes (will not be taken into account)
 
-GIT_STATUS="$(git status --porcelain -- $(sed 's/^./:!&/' <<< $EXTERNAL_PLUGINS))"
-if [ "" != "$GIT_STATUS" -a "$USE_STASH" != "yes" ]; then
+# We do want word splitting, disable SC2046
+# shellcheck disable=SC2046
+GIT_STATUS="$(git status --porcelain -- $(printf ":!%s\n" "${EXTERNAL_PLUGINS[@]}"))"
+if [ "" != "$GIT_STATUS" ] && [ "$USE_STASH" != "yes" ]; then
   echo "WARNING: uncommitted changes will be IGNORED when making archive:"
+  # Unable to replace sed with ${var//search/replace} here, disable SC2001
+  # shellcheck disable=SC2001
   echo "$GIT_STATUS" | sed 's/^/  /'
   echo "----------------------------------------------------------------"
 fi
@@ -151,26 +178,26 @@ if [ "$USE_STASH" == "yes" ]; then
     ARCHIVE_COMMIT=$(git stash create)
 fi
 
-git archive ${ARCHIVE_COMMIT:-HEAD} -o $FRAMAC_TAR --prefix "$FRAMAC/"
+git archive "${ARCHIVE_COMMIT:-HEAD}" -o "$FRAMAC_TAR" --prefix "$FRAMAC/"
 
 ################################################################################
-# Add external plugin to archive
+# Add external plugins to archive
 
-if [ "" != "$EXTERNAL_PLUGINS" ]
+if [ "${#EXTERNAL_PLUGINS[@]}" -gt 0 ]
 then
   echo "Including external plugins:"
 fi
 
-for plugin in $EXTERNAL_PLUGINS
+for plugin in "${EXTERNAL_PLUGINS[@]}"
 do
     echo "  $plugin"
-    PLUGIN_TAR="$(basename $plugin).tar"
-    git -C $plugin archive HEAD -o $PLUGIN_TAR --prefix "$FRAMAC/$plugin/"
-    $TAR --concatenate --file=$FRAMAC_TAR "$plugin/$PLUGIN_TAR"
-    rm -rf "$plugin/$PLUGIN_TAR"
+    PLUGIN_TAR="$(basename "$plugin").tar"
+    git -C "$plugin" archive HEAD -o "$PLUGIN_TAR" --prefix "$FRAMAC/$plugin/"
+    $TAR --concatenate --file="$FRAMAC_TAR" "$plugin/$PLUGIN_TAR"
+    rm -rf "${plugin:?}/${PLUGIN_TAR:?}"
 done
 
-if [ "" != "$EXTERNAL_PLUGINS" ]
+if [ "${#EXTERNAL_PLUGINS[@]}" -gt 0 ]
 then
   echo "----------------------------------------------------------------"
 fi
@@ -186,14 +213,14 @@ HEADER_SPEC="header-spec.txt"
 git ls-files |\
 git check-attr --stdin export-ignore |\
 grep -v "export-ignore: set" | awk -F ': ' '{print $1}' |\
-git check-attr --stdin header_spec > $HEADER_SPEC
+git check-attr --stdin header_spec > "$HEADER_SPEC"
 
-for plugin in $EXTERNAL_PLUGINS ; do
-  git -C $plugin ls-files |\
-  git -C $plugin check-attr --stdin export-ignore |\
+for plugin in "${EXTERNAL_PLUGINS[@]}" ; do
+  git -C "$plugin" ls-files |\
+  git -C "$plugin" check-attr --stdin export-ignore |\
   grep -v "export-ignore: set" | awk -F ': ' '{print $1}' |\
-  git -C $plugin check-attr --stdin header_spec |\
-  xargs -n3 printf "$plugin/%s %s %s\n" >> $HEADER_SPEC
+  git -C "$plugin" check-attr --stdin header_spec |\
+  xargs -n3 printf "$plugin/%s %s %s\n" >> "$HEADER_SPEC"
 done
 
 ################################################################################
@@ -203,7 +230,7 @@ CHECK_HEADER_OPT="-header-dirs headers"
 
 # For plugins, either they can be open-source and we assume they have OS headers
 # or they are closed-source
-for plugin in $PLUGINS ; do
+for plugin in "${PLUGINS[@]}" ; do
   if [ -d "$plugin/headers" ] ; then
     CHECK_HEADER_OPT="$CHECK_HEADER_OPT -header-dirs $plugin/headers"
   fi
@@ -215,9 +242,11 @@ done
 echo "Check headers..."
 
 TMP_DIR=$(mktemp -d)
-$TAR xf $FRAMAC_TAR -C $TMP_DIR
+$TAR xf "$FRAMAC_TAR" -C "$TMP_DIR"
 
 # Check
+# We do want globbing and word splitting, disable SC2086
+# shellcheck disable=SC2086
 $HDRCK $CHECK_HEADER_OPT -spec-format="3-fields-by-line" -C "$TMP_DIR/$FRAMAC" $HEADER_SPEC
 
 ################################################################################
@@ -225,25 +254,25 @@ $HDRCK $CHECK_HEADER_OPT -spec-format="3-fields-by-line" -C "$TMP_DIR/$FRAMAC" $
 
 echo "Finalizing archive..."
 
-echo $VERSION_SAFE > $TMP_DIR/$FRAMAC/VERSION
-echo $VERSION_CODENAME > $TMP_DIR/$FRAMAC/VERSION_CODENAME
+echo "$VERSION_SAFE" > "$TMP_DIR/$FRAMAC/VERSION"
+echo "$VERSION_CODENAME" > "$TMP_DIR/$FRAMAC/VERSION_CODENAME"
 
 DATE="$(date +%F)"
 
-$TAR czf $FRAMAC_TAR.gz -C $TMP_DIR $FRAMAC \
+$TAR czf "$FRAMAC_TAR.gz" -C "$TMP_DIR" "$FRAMAC" \
   --numeric-owner --owner=0 --group=0 --sort=name --mode='a+rw' \
   --mtime="$DATE Z"
 
 if [[ "$CI_LINK" == "yes" ]]; then
-  ln $FRAMAC_TAR.gz "frama-c.tar.gz"
+  ln "$FRAMAC_TAR.gz" "frama-c.tar.gz"
 fi
 
 ################################################################################
 # Cleaning
 
-rm -rf $HEADER_SPEC
-rm -rf $FRAMAC_TAR
-rm -rf $TMP_DIR
+rm -rf "$HEADER_SPEC"
+rm -rf "$FRAMAC_TAR"
+rm -rf "$TMP_DIR"
 
 echo "Generated: $FRAMAC_TAR.gz"
 echo "----------------------------------------------------------------"
